feat: 扩大实验范围 — 新数据源 + 多步时距 + 点位回归任务
- 数据范围：新增出口年率（外需）与沪深300月收益/月内波动（市场，零发布滞后），
  特征 26→35；修复同月多报告行时空预告行覆盖有值行的去重缺陷
- 多步时距实验 horizon.py：h=1/2/3 月荣枯预测，purge 随 h 扩大防标签前视，
  产出 AUC 衰减曲线
- 第三任务 level_task.py：PMI 点位回归（持续/Ridge/RF/XGB），
  MAE/RMSE/隐含荣枯 + 配对自助 Diebold-Mariano 检验
- 标签按连续日历索引 shift，漏发月份不再错位；消融新增外需/市场两组
- 新增 14 项单测（总 55），修复全部 ruff 0.16 lint

+++ src/pmi_nowcast/models.py
@@ -29,7 +29,7 @@ class NaiveMajority:
 
     def fit(self, X, y):
         y = np.asarray(y)
-        self.majority_ = int(round(y.mean()))
+        self.majority_ = round(float(y.mean()))
         self.p_ = float(y.mean())
         return self
